Change heading ids to kramdown style
The anchor links which kramdown generates in the table of contents are
based on the heading's HTML content and add an id to the header itself.
Since it uses the HTML content, this duplicated the anchor into the toc.
Links in the docs which referred to these anchors would then go to the
table of contents instead of the heading itself.

For example, the headings as they are now generate in the table of
contents to look like:

```html
<li>
  <a href="#features-of-cedar" id="markdown-toc-features-of-cedar">Features of Cedar</a>
  <a name="feature-overview"></a>
</li>
```

While the heading is:

```html
<h2 id="features-of-cedar">
<!-- Content omitted -->
Features of Cedar<a name="feature-overview"></a>
</h2>
```

This change specifies the anchors manually to prevent kramdown creating
new ones -- preventing the double destination problem.