Reduce confusion when looking at action groups example. (#153)
When reading through the examples, the one on action groups is structured differently to the others, with the action filter done in the `when` condition instead of the scope. I have seen this lead to confusion as to whether it was possible to use action groups in scopes. This PR changes the example to use the scope which reduces that confusion and makes the example more consistent with the other action examples.

Signed-off-by: Ben Edwards <ben.edwards@benjaminsedwards.com>